move to package st.extreme.math.fraction
