+++ src/main/java/st/extreme/math/fraction/BigFraction.java
@@ -1,4 +1,4 @@
-package st.extreme.math;
+package st.extreme.math.fraction;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
