+++ src/main/java/st/extreme/math/fraction/package-info.java
@@ -2,4 +2,4 @@
 /**
  * Provides the classes to perform exact calculations with fractions, also known as rational numbers.
  */
-package st.extreme.math;
+package st.extreme.math.fraction;
