+++ src/test/java/st/extreme/math/fraction/BigFractionPatternTest.java
@@ -1,9 +1,11 @@
-package st.extreme.math;
+package st.extreme.math.fraction;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
+
+import st.extreme.math.fraction.BigFraction;
 
 public class BigFractionPatternTest {
