+++ src/test/java/st/extreme/math/fraction/BigFractionRepresentationTest.java
@@ -1,9 +1,11 @@
-package st.extreme.math;
+package st.extreme.math.fraction;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
+
+import st.extreme.math.fraction.BigFraction;
 
 public class BigFractionRepresentationTest {
